lesson12

+++ API/src/main/java/ru/geekbrains/web/api/dtos/CartDto.java
@@ -1,10 +1,11 @@
 package ru.geekbrains.web.api.dtos;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CartDto {
     private List<OrderItemDTO> items;
-    private int totalPrice;
+    private BigDecimal totalPrice;
 
     public List<OrderItemDTO> getItems() {
         return items;
@@ -14,18 +15,18 @@ public class CartDto {
         this.items = items;
     }
 
-    public int getTotalPrice() {
+    public BigDecimal getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(BigDecimal totalPrice) {
         this.totalPrice = totalPrice;
     }
 
     public CartDto() {
     }
 
-    public CartDto(List<OrderItemDTO> items, int totalPrice) {
+    public CartDto(List<OrderItemDTO> items, BigDecimal totalPrice) {
         this.items = items;
         this.totalPrice = totalPrice;
     }
